split stories
