+++ packages/react-badge/src/stories/CounterBadgeDefault.stories.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+// https://github.com/microsoft/fluentui/pull/18695#issuecomment-868432982
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
+import { CounterBadge, CounterBadgeProps } from '@fluentui/react-badge';
+
+export const Default = (args: CounterBadgeProps) => <CounterBadge {...args} />;
+Default.args = {
+  count: 5,
+};
