+++ packages/react-badge/src/stories/PresenceBadgeDefault.stories.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+// https://github.com/microsoft/fluentui/pull/18695#issuecomment-868432982
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
+import { PresenceBadge } from '@fluentui/react-badge';
+
+export const Default = () => <PresenceBadge />;
